agent/pyproject.toml
[project]
name = "scout-voice-agent"
version = "0.1.0"
description = "Voice-enabled broker assistant using Strands SDK with Nova Sonic and AgentCore Gateway"
requires-python = ">=3.11"
dependencies = [
    "strands-agents[bidi-all]>=1.18.0",
    "strands-agents-tools>=0.2.0",
    "mcp>=1.0.0",
    "httpx-auth-awssigv4>=0.1.0",
    "boto3>=1.41.0",
    "bedrock-agentcore>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

